Configurable deflines #13574
