+++ Client/src/components/reporters/BedGeneReporterForm.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { RadioList, CheckboxList, SingleSelect, TextBox, Checkbox } from '@veupathdb/wdk-client/lib/Components';
+import { deflineFieldOptions, GenomicSequenceRegionInputs, ProteinSequenceRegionInputs, FeaturesList, ComponentsList } from './BedFormElements';
 import * as ComponentUtils from '@veupathdb/wdk-client/lib/Utils/ComponentUtils';
 import * as ReporterUtils from '@veupathdb/wdk-client/lib/Views/ReporterForm/reporterUtils';
 import SrtHelp from '../SrtHelp';
@@ -7,13 +8,9 @@ import SrtHelp from '../SrtHelp';
 let util = Object.assign({}, ComponentUtils, ReporterUtils);
 
 
-let deflineFieldOptions = [
-  { value: 'organism', display: 'Organism' },
-  { value: 'description', display: 'Description' },
-  { value: 'position', display: 'Position' },
-  { value: 'ui_choice', display: 'Query' },
-  { value: 'segment_length', display: 'Length' },
-
+let splicedGenomicOptions = [
+  { value: 'cds', display: 'CDS'},
+  { value: 'transcript', display: 'Transcript'},
 ];
 
 let proteinFeatureOptions = [
@@ -30,129 +27,6 @@ let geneComponentOptions = [
   { value: 'intron', display: 'intron' },
 ];
 
-let genomicAnchorValues = [
-  { value: 'Start', display: 'Transcription Start***' },
-  { value: 'CodeStart', display: 'Translation Start (ATG)' },
-  { value: 'CodeEnd', display: 'Translation Stop Codon' },
-  { value: 'End', display: 'Transcription Stop***' }
-];
-
-let proteinAnchorValues = [
-  { value: 'Start', display: 'Downstream from Start' },
-  { value: 'End', display: 'Upstream from End' }
-];
-
-let signs = [
-  { value: 'plus', display: '+' },
-  { value: 'minus', display: '-' }
-];
-
-let SequenceRegionRange = props => {
-  let { label, anchor, sign, offset, formState, getUpdateHandler } = props;
-  return (
-    <React.Fragment>
-      <span>{label}</span>
-      <SingleSelect name={anchor} value={formState[anchor]}
-          onChange={getUpdateHandler(anchor)} items={genomicAnchorValues}/>
-      <SingleSelect name={sign} value={formState[sign]}
-          onChange={getUpdateHandler(sign)} items={signs}/>
-      <TextBox name={offset} value={formState[offset]}
-          onChange={getUpdateHandler(offset)} size="6"/>
-      nucleotides
-    </React.Fragment>
-  );
-};
-
-let ProteinRegionRange = props => {
-  let { label, anchor, offset, formState, getUpdateHandler } = props;
-  return (
-    <React.Fragment>
-      <span>{label}</span>
-      <SingleSelect name={anchor} value={formState[anchor]}
-          onChange={getUpdateHandler(anchor)} items={proteinAnchorValues}/>
-      <TextBox name={offset} value={formState[offset]}
-          onChange={getUpdateHandler(offset)} size="6"/>
-      aminoacids
-    </React.Fragment>
-  );
-};
-
-let GenomicSequenceRegionInputs = props => {
-  let { formState, getUpdateHandler } = props;
-  return (
-    <div>
-      <div style={{marginLeft:"0.75em"}}>
-        <Checkbox name="reverseAndComplement" value={formState.reverseAndComplement} onChange={getUpdateHandler('reverseAndComplement')}/> Reverse & Complement
-      </div>
-      <div
-        style={{
-          display: 'inline-grid',
-          gridTemplateColumns: 'repeat(5, auto)',
-          alignItems: 'center',
-          gridRowGap: '0.25em',
-          gridColumnGap: '0.5em',
-          marginLeft: '0.75em'
-        }}
-      >
-        <SequenceRegionRange label="Begin at" anchor="upstreamAnchor" sign="upstreamSign"
-          offset="upstreamOffset" formState={formState} getUpdateHandler={getUpdateHandler}/>
-        <SequenceRegionRange label="End at" anchor="downstreamAnchor" sign="downstreamSign"
-          offset="downstreamOffset" formState={formState} getUpdateHandler={getUpdateHandler}/>
-      </div>
-    </div>
-  );
-}
-let ProteinSequenceRegionInputs = props => {
-  let { formState, getUpdateHandler } = props;
-  return (
-    <div>
-      <div
-        style={{
-          display: 'inline-grid',
-          gridTemplateColumns: 'repeat(4, auto)',
-          alignItems: 'center',
-          gridRowGap: '0.25em',
-          gridColumnGap: '0.5em',
-          marginLeft: '0.75em'
-        }}
-      >
-        <ProteinRegionRange label="Begin at" anchor="startAnchor3" offset="startOffset3"
-          formState={formState} getUpdateHandler={getUpdateHandler}/>
-        <ProteinRegionRange label="End at" anchor="endAnchor3" offset="endOffset3"
-          formState={formState} getUpdateHandler={getUpdateHandler}/>
-      </div>
-    </div>
-  );
-};
-
-let FeaturesList = props => {
-  let { features, field, formState, getUpdateHandler} = props;
-  return (
-    <div>
-      <div
-        style={{
-          marginLeft: '0.75em'
-        }}>
-        <RadioList name={field} value={formState[field]} onChange={getUpdateHandler(field)} items={features}/>
-      </div>
-    </div>
-  );
-}
-
-// RadioList -> multiple choice
-let ComponentsList = props => {
-  let { features, field, formState, getUpdateHandler} = props;
-  return (
-    <div>
-      <div
-        style={{
-          marginLeft: '0.75em'
-        }}>
-        <CheckboxList name={field} value={formState[field]} onChange={getUpdateHandler(field)} items={features} linksPosition={null}/>
-      </div>
-    </div>
-  );
-}
 
 /** @type import('./Types').ReporterFormComponent */
 let FastaGeneReporterForm = props => {
@@ -168,12 +42,21 @@ let FastaGeneReporterForm = props => {
         <RadioList name="type" value={formState.type}
           onChange={typeUpdateHandler} items={
           [
-            { value: 'genomic', display: 'Genomic Sequence', body:  (<GenomicSequenceRegionInputs formState={formState} getUpdateHandler={getUpdateHandler}/>)},
-            { value: 'protein', display: 'Protein Sequence', body: (<ProteinSequenceRegionInputs formState={formState} getUpdateHandler={getUpdateHandler}/>)},
-            { value: 'protein_features', display: 'Protein Features', body: (<FeaturesList field="proteinFeature" features={proteinFeatureOptions} formState={formState} getUpdateHandler={getUpdateHandler} />) },
-            { value: 'cds', display: 'CDS'},
-            { value: 'transcript', display: 'Transcript'},
-            { value: 'gene_components', display: 'Gene Components', body: (<ComponentsList field="geneComponents" features={geneComponentOptions} formState={formState} getUpdateHandler={getUpdateHandler} />) },
+            { value: 'genomic', display: 'Unspliced Genomic Sequence', body:  (
+              <GenomicSequenceRegionInputs formState={formState} getUpdateHandler={getUpdateHandler}/>
+            )},
+            { value: 'protein', display: 'Protein Sequence', body: (
+              <ProteinSequenceRegionInputs formState={formState} getUpdateHandler={getUpdateHandler}/>
+            )},
+            { value: 'protein_features', display: 'Protein Features', body: (
+              <FeaturesList field="proteinFeature" features={proteinFeatureOptions} formState={formState} getUpdateHandler={getUpdateHandler} />
+            )},
+            { value: 'spliced_genomic', display: 'Spliced Genomic Sequence', body: (
+              <FeaturesList field="splicedGenomic" features={splicedGenomicOptions} formState={formState} getUpdateHandler={getUpdateHandler} />
+            )},
+            { value: 'gene_components', display: 'Gene Components', body: (
+              <ComponentsList field="geneComponents" features={geneComponentOptions} formState={formState} getUpdateHandler={getUpdateHandler} />
+            )},
           ]  
           
           }/>
@@ -187,8 +70,11 @@ let FastaGeneReporterForm = props => {
       <div style={{marginLeft:"2em"}}>
         <RadioList name="deflineType" value={formState.deflineType}
           onChange={getUpdateHandler('deflineType')} items={[
-              {  value: "short", display: 'ID Only' },
-            {  value: "full", display: 'Full Fasta Header', body: (<RadioList name="deflineFields" value={formState.deflineType} onChange={getUpdateHandler('deflineType')}/> items={deflineFieldOptions}) }
+            {  value: "short", display: 'ID Only' },
+            { 
+              value: "full", display: 'Full Fasta Header',
+              body: (<ComponentsList field="deflineFields" features={deflineFieldOptions} formState={formState} getUpdateHandler={getUpdateHandler} />),
+            }
             ]}/>
       </div>
       { includeSubmit &&
@@ -217,7 +103,7 @@ FastaGeneReporterForm.getInitialState = () => ({
     attachmentType: 'plain',
     type: 'genomic',
     reverseAndComplement: false,
-    deflineType: deflineTypes[0].value,
+    deflineType: "short",
     deflineFields: deflineFieldOptions.map((x) => x.value),
 
     // sequence region inputs for 'genomic'
@@ -235,6 +121,7 @@ FastaGeneReporterForm.getInitialState = () => ({
     endOffset3: 0,
     geneComponents: geneComponentOptions.map((x) => x.value),
     proteinFeature: proteinFeatureOptions[0].value,
+    splicedGenomic: splicedGenomicOptions[0].value,
   },
   formUiState: {}
 });
