+++ Client/src/components/reporters/BedSequenceReporterForm.jsx
@@ -1,15 +1,11 @@
 import React from 'react';
 import { RadioList } from '@veupathdb/wdk-client/lib/Components';
+import { deflineFieldOptions, ComponentsList } from './BedFormElements';
 import * as ComponentUtils from '@veupathdb/wdk-client/lib/Utils/ComponentUtils';
 import * as ReporterUtils from '@veupathdb/wdk-client/lib/Views/ReporterForm/reporterUtils';
 
 
 let util = Object.assign({}, ComponentUtils, ReporterUtils);
-
-let deflineTypes = [
-  {  value: "short", display: 'ID Only' },
-  {  value: "full", display: 'Full Fasta Header' }
-];
 
 /** @type import('./Types').ReporterFormComponent */
 const BedSegmentableSequenceReporterForm = props => {
@@ -26,7 +22,13 @@ const BedSegmentableSequenceReporterForm = props => {
       <h3>Fasta defline:</h3>
       <div style={{marginLeft:"2em"}}>
         <RadioList name="deflineType" value={formState.deflineType}
-          onChange={getUpdateHandler('deflineType')} items={deflineTypes}/>
+          onChange={getUpdateHandler('deflineType')} items={[
+            {  value: "short", display: 'ID Only' },
+            { 
+              value: "full", display: 'Full Fasta Header',
+              body: (<ComponentsList field="deflineFields" features={deflineFieldOptions} formState={formState} getUpdateHandler={getUpdateHandler} />),
+            }
+            ]}/>
       </div>
       { includeSubmit &&
         <div style={{margin:'0.8em'}}>
@@ -40,7 +42,8 @@ const BedSegmentableSequenceReporterForm = props => {
 BedSegmentableSequenceReporterForm.getInitialState = () => ({
   formState: {
     attachmentType: 'plain',
-    deflineType: deflineTypes[0].value,
+    deflineType: "short",
+    deflineFields: deflineFieldOptions.map((x) => x.value),
   },
   formUiState: {}
 });
